Added SudokuCell file for the class.
